Improve usage of constraint_impulse, move update_inertias to its own file,  fix issues with destroyed constraints not being removed from the island contexts.
