+++ CMakeLists.txt
@@ -78,6 +78,7 @@ set(Edyn_SOURCES
     src/edyn/sys/apply_gravity.cpp
     src/edyn/sys/update_aabbs.cpp
     src/edyn/sys/update_rotated_meshes.cpp
+    src/edyn/sys/update_inertias.cpp
     src/edyn/util/rigidbody.cpp
     src/edyn/util/constraint_util.cpp
     src/edyn/util/shape_util.cpp
