+++ include/edyn/serialization/constraint/hinge_constraint_s11n.hpp
@@ -8,7 +8,6 @@ namespace edyn {
 template<typename Archive>
 void serialize(Archive &archive, hinge_constraint &c) {
     archive(c.body);
-    archive(c.impulse);
     archive(c.pivot);
     archive(c.frame);
 }
