+++ include/edyn/serialization/constraint/soft_distance_constraint_s11n.hpp
@@ -8,7 +8,6 @@ namespace edyn {
 template<typename Archive>
 void serialize(Archive &archive, soft_distance_constraint &c) {
     archive(c.body);
-    archive(c.impulse);
     archive(c.pivot);
     archive(c.distance);
     archive(c.stiffness);
