+++ src/edyn/constraints/contact_constraint.cpp
@@ -10,6 +10,7 @@
 #include "edyn/comp/mass.hpp"
 #include "edyn/comp/inertia.hpp"
 #include "edyn/collision/contact_point.hpp"
+#include "edyn/constraints/constraint_impulse.hpp"
 #include "edyn/dynamics/row_cache.hpp"
 #include "edyn/util/constraint_util.hpp"
 #include <entt/entt.hpp>
@@ -23,10 +24,11 @@ struct row_start_index_contact_constraint {
 void prepare_contact_constraints(entt::registry &registry, row_cache &cache, scalar dt) {
     auto body_view = registry.view<position, orientation, linvel, angvel, mass_inv, inertia_world_inv, delta_linvel, delta_angvel>();
     auto con_view = registry.view<contact_constraint, contact_point>();
+    auto imp_view = registry.view<constraint_impulse>();
     size_t row_idx = cache.con_rows.size();
     registry.ctx_or_set<row_start_index_contact_constraint>(row_idx);
 
-    con_view.each([&] (contact_constraint &con, contact_point &cp) {
+    con_view.each([&] (entt::entity entity, contact_constraint &con, contact_point &cp) {
         auto [posA, ornA, linvelA, angvelA, inv_mA, inv_IA, dvA, dwA] = body_view.get<position, orientation, linvel, angvel, mass_inv, inertia_world_inv, delta_linvel, delta_angvel>(con.body[0]);
         auto [posB, ornB, linvelB, angvelB, inv_mB, inv_IB, dvB, dwB] = body_view.get<position, orientation, linvel, angvel, mass_inv, inertia_world_inv, delta_linvel, delta_angvel>(con.body[1]);
 
@@ -83,6 +85,7 @@ void prepare_contact_constraints(entt::registry &registry, row_cache &cache, sca
         
         auto num_rows = 2;
         auto options = std::array<constraint_row_options, 2>{normal_options, {}};
+        auto &imp = imp_view.get(entity);
 
         for (size_t i = 0; i < num_rows; ++i) {
             auto j = i + row_idx;
@@ -98,7 +101,7 @@ void prepare_contact_constraints(entt::registry &registry, row_cache &cache, sca
             row.dwA = &dwA;
             row.dwB = &dwB;
 
-            row.impulse = con.impulse[i];
+            row.impulse = imp.values[i];
 
             prepare_row(row, options[i], linvelA, linvelB, angvelA, angvelB);
             warm_start(row);
