+++ src/edyn/constraints/generic_constraint.cpp
@@ -1,6 +1,6 @@
 #include "edyn/constraints/generic_constraint.hpp"
 #include "edyn/comp/constraint_row.hpp"
-#include "edyn/constraints/constraint.hpp"
+#include "edyn/constraints/constraint_impulse.hpp"
 #include "edyn/math/constants.hpp"
 #include "edyn/math/matrix3x3.hpp"
 #include "edyn/comp/position.hpp"
@@ -23,6 +23,7 @@ void prepare_generic_constraints(entt::registry &registry, row_cache &cache, sca
                                    mass_inv, inertia_world_inv, 
                                    delta_linvel, delta_angvel>();
     auto con_view = registry.view<generic_constraint, constraint_impulse>();
+
     con_view.each([&] (generic_constraint &con, constraint_impulse &imp) {
         auto [posA, ornA, linvelA, angvelA, inv_mA, inv_IA, dvA, dwA] = 
             body_view.get<position, orientation, linvel, angvel, mass_inv, inertia_world_inv, delta_linvel, delta_angvel>(con.body[0]);
