+++ src/edyn/constraints/hinge_constraint.cpp
@@ -1,5 +1,5 @@
 #include "edyn/constraints/hinge_constraint.hpp"
-#include "edyn/constraints/constraint.hpp"
+#include "edyn/constraints/constraint_impulse.hpp"
 #include "edyn/math/constants.hpp"
 #include "edyn/math/matrix3x3.hpp"
 #include "edyn/comp/position.hpp"
@@ -48,6 +48,7 @@ void prepare_hinge_constraints(entt::registry &registry, row_cache &cache, scala
                                    mass_inv, inertia_world_inv, 
                                    delta_linvel, delta_angvel>();
     auto con_view = registry.view<hinge_constraint, constraint_impulse>();
+
     con_view.each([&] (hinge_constraint &con, constraint_impulse &imp) {
         auto [posA, ornA, linvelA, angvelA, inv_mA, inv_IA, dvA, dwA] = 
             body_view.get<position, orientation, linvel, angvel, mass_inv, inertia_world_inv, delta_linvel, delta_angvel>(con.body[0]);
